Updating to thrust 1.9.6: macros, namespaces, algorithms with new api

+++ hydra/detail/external/thrust/pair.h
@@ -228,7 +228,7 @@ template <typename T1, typename T2>
  *  \tparam N This parameter selects the member of interest.
  *  \tparam T A \c pair type of interest.
  */
-template<int N, typename T> struct tuple_element;
+template<long unsigned  N, typename T> struct tuple_element;
 
 
 /*! This convenience metafunction is included for compatibility with
